refactor: Move /addkit command and remove /editkit
This commit refactors the kit commands for better organization:

- The logic for the `/addkit` command has been moved from `addkit.js` to `kit.js`.
- The now-redundant `addkit.js` file has been deleted.
- The `/editkit` command and its corresponding file `editkit.js` have been removed, as the kit editor is now opened automatically by `/addkit`.
