+++ AddonExeBP/scripts/modules/commands/kit.js
@@ -3,9 +3,10 @@ import { commandManager } from './commandManager.js';
 import * as kitsManager from '../../core/kitsManager.js';
 import { getConfig } from '../../core/configManager.js';
 import { errorLog } from '../../core/errorLogger.js';
-import { createKit } from '../../core/kitAdminManager.js';
+import { createKit, getAllKits } from '../../core/kitAdminManager.js';
 import { addItemToKit } from '../../core/kitItemsManager.js';
 import { formatCooldown } from '../../core/utils.js';
+import { showPanel } from '../../core/uiManager.js';
 
 commandManager.register({
     name: 'kit',
@@ -74,18 +75,27 @@ commandManager.register({
 
 commandManager.register({
     name: 'addkit',
-    description: 'Create a new kit from your inventory.',
+    description: 'Create a new kit from your inventory and open the editor.',
     permissionLevel: 1, // Admins only
     allowConsole: false,
     parameters: [
-        { name: 'kitName', type: 'string', description: 'The name of the kit to create.' }
+        { name: 'kitName', type: 'string', description: 'The name for the new kit. Leave blank to auto-generate.', optional: true }
     ],
     execute: (player, args) => {
-        const { kitName } = args;
+        let { kitName } = args;
+
+        if (!kitName) {
+            const allKits = getAllKits();
+            let i = 1;
+            kitName = 'kit';
+            while (allKits[kitName]) {
+                i++;
+                kitName = `kit${i}`;
+            }
+        }
 
         const inventory = player.getComponent('minecraft:inventory').container;
         const items = [];
-        // Player inventory is slots 0-35.
         for (let i = 0; i < 36; i++) {
             const item = inventory.getItem(i);
             if (item) {
@@ -107,10 +117,12 @@ commandManager.register({
             return player.sendMessage(`§c${createResult.message}`);
         }
 
+        const lowerCaseKitName = kitName.toLowerCase();
         for (const item of items) {
-            addItemToKit(kitName, item);
+            addItemToKit(lowerCaseKitName, item);
         }
 
-        player.sendMessage(`§aSuccessfully created kit '${kitName}' with ${items.length} item stacks.`);
+        player.sendMessage(`§aSuccessfully created kit '${lowerCaseKitName}'. Opening editor...`);
+        showPanel(player, `kitActionMenu_${lowerCaseKitName}`);
     }
 });
